# systemtest/loadgen: don't pre-compress batches