## Motivation/summary

Compress batches while sending, rather than ahead of time. This will enable us to transform events (https://github.com/elastic/apm-server/issues/10445 and https://github.com/elastic/apm-server/issues/10446) and may also simplify state-state event generation (https://github.com/elastic/apm-server/pull/10468#pullrequestreview-1348434618).

This changes does add significant runtime overhead to the load generator, but:
 - we mitigate that somewhat by using `zlib.BestSpeed`, and using the more optimised `klauspost/compress` package
 - `apmbench` is ultimately limited by APM Server's throughput, and the difference doesn't impact real world benchmarks. I ran `testing/benchmark` twice, once on main and once with `apmbench` built off this branch, with `BENCHMARK_COUNT=6 BENCHMARK_RUN=BenchmarkAgentAll`. There were no significant differences:

```
            │ /tmp/old.txt │        /tmp/new.txt         │
            │    sec/op    │   sec/op    vs base         │
AgentAll-64     12.39 ± 1%   12.42 ± 2%  ~ (p=0.485 n=6)

            │    /tmp/old.txt     │              /tmp/new.txt              │
            │ error_responses/sec │ error_responses/sec  vs base           │
AgentAll-64            0.000 ± 0%            0.000 ± 0%  ~ (p=1.000 n=6) ¹
¹ all samples are equal

            │ /tmp/old.txt │        /tmp/new.txt         │
            │  errors/sec  │ errors/sec  vs base         │
AgentAll-64     29.05 ± 1%   28.99 ± 2%  ~ (p=0.513 n=6)

            │ /tmp/old.txt │         /tmp/new.txt         │
            │  events/sec  │ events/sec   vs base         │
AgentAll-64    1.463k ± 2%   1.458k ± 2%  ~ (p=0.372 n=6)

            │ /tmp/old.txt │        /tmp/new.txt         │
            │  gc_cycles   │ gc_cycles   vs base         │
AgentAll-64     57.00 ± 7%   57.00 ± 5%  ~ (p=0.818 n=6)

            │  /tmp/old.txt  │          /tmp/new.txt           │
            │ max_goroutines │ max_goroutines  vs base         │
AgentAll-64       57.00 ± 2%       58.00 ± 9%  ~ (p=0.431 n=6)

            │  /tmp/old.txt  │          /tmp/new.txt           │
            │ max_heap_alloc │ max_heap_alloc  vs base         │
AgentAll-64      311.2M ± 6%      300.8M ± 4%  ~ (p=0.589 n=6)

            │   /tmp/old.txt   │           /tmp/new.txt            │
            │ max_heap_objects │ max_heap_objects  vs base         │
AgentAll-64       2.512M ± 11%       2.340M ± 10%  ~ (p=0.310 n=6)

            │ /tmp/old.txt │         /tmp/new.txt         │
            │   max_rss    │   max_rss    vs base         │
AgentAll-64    363.8M ± 4%   368.4M ± 4%  ~ (p=0.699 n=6)

            │      /tmp/old.txt       │               /tmp/new.txt               │
            │ mean_available_indexers │ mean_available_indexers  vs base         │
AgentAll-64                9.139 ± 1%                9.136 ± 1%  ~ (p=0.675 n=6)

            │ /tmp/old.txt │         /tmp/new.txt         │
            │ metrics/sec  │ metrics/sec  vs base         │
AgentAll-64     230.6 ± 2%    229.1 ± 2%  ~ (p=0.394 n=6)

            │ /tmp/old.txt │        /tmp/new.txt         │
            │  spans/sec   │ spans/sec   vs base         │
AgentAll-64     871.7 ± 1%   869.8 ± 2%  ~ (p=0.485 n=6)

            │ /tmp/old.txt │        /tmp/new.txt         │
            │   txs/sec    │  txs/sec    vs base         │
AgentAll-64     331.4 ± 1%   330.8 ± 2%  ~ (p=0.513 n=6)

            │ /tmp/old.txt │         /tmp/new.txt          │
            │     B/op     │     B/op      vs base         │
AgentAll-64   554.4Mi ± 0%   553.8Mi ± 0%  ~ (p=0.485 n=6)

            │ /tmp/old.txt │           /tmp/new.txt            │
            │  allocs/op   │  allocs/op   vs base              │
AgentAll-64    7.986M ± 0%   7.994M ± 0%  +0.10% (p=0.004 n=6)
```

## Checklist

~- [ ] Update [CHANGELOG.asciidoc](https://github.com/elastic/apm-server/blob/main/CHANGELOG.asciidoc)~
~- [ ] Update [package changelog.yml](https://github.com/elastic/apm-server/blob/main/apmpackage/apm/changelog.yml) (only if changes to `apmpackage` have been made)~
~- [ ] Documentation has been updated~

## How to test these changes

<!--
Explain how this PR can be tested by the reviewer: commands, dependencies, steps, etc.
If it is self-explanatory, delete this section.
-->

## Related issues

<!--
Reference the related issue(s), and make use of magic keywords where it makes sense
https://help.github.com/articles/closing-issues-using-keywords/.
-->